pr fixes
